No prange and test_minim_algo.py works
The resource "computation.pyx" is such that it does not utilise
cython.prange. With this setting, the script "test_minim_algo.py"
reveals that the class MinimisationProcedure in the resource
"minimisation_algo.pyx" is capable of performing gradient descent in
parallel using MultiProcessing.Pool. In this commit, the function "compute_f_and_grad" of
"minimisation_algo.pyx" is exterior to the function
"gradient_descent_partial"