Simpler random update and parallel works
In the resource "minimisation_algo.pyx" the function
"gradient_descent_partial" was such that when the descent along the
gradient failed to produce a better point, it updated the position using
a multivariate normal move from the previous position. The script
"test_minim_algo_2.py" revealed that this was one of the causes for the
parallel gradient descent to break. Hence, the following modification
was implemented.
Instead of using the multivariate normal, the random updated is an
infinitesimal step along the direction of the gradient, where the lenght
of this infinitesimal step is random, produced using the random number generator
"rand" of libc.stdlib.
The script "test_minim_algo_2.py" shows that after such modification the
class EstimProcedure of the resource "mle_estimation.pyx" is
capable of successfully launching the parallel gradient descent
implemented in "minimisation_algo.pyx".