test_mle_estim successful!
Leveraging on the corrections to the resource "minimisation_algo.pyx"
already committed in the previous version, this commit incorporates
small tunings to let the class EstimProcedure in "mle_estimation.pyx" to
be callable from "model.pyx" and to be able to perform parallel gradient
descent both for the estimation of the state-depenedent hawkes process
and for the pre-estimation of the ordinary hawkes process. This is
demonstrated in tests/test_mle_estim.py.