EstimProcedure(s) fully incorporated in model.pyx
The resources "mle_estimation.pyx" and "nonparam_estimation.pyx" contain
classes EstimProcedure. Their usage was not fully exploited by
"model.pyx", with particular regards to the way in which it handled the
partial calibrations of each component. With this commit, "model.pyx" is
written to rely on those two EStimProcedure(s) as much as possible,
unloading self from task and passing those to the resources.