Update of calibration/main_EliteBook.py
The python script "main_EliteBook.py" in the directiory "calibration" is
the one to handle the calibration of the model when this is performed on
my laptop HP_EliteBook. Previously, it did not use the most recent
implementations in the resources "minimisation_algo.pyx",
"mle_estimation.pyx", or "nonparam_estimation.pyx". This commit saves
the cahnges to main_EliteBook.py written in order to use the most recent
resources.