Bookmark: test_estim_model_2020-04-03_10134
Towards the final test of the various estimation procedures to present
in the paper. The python script test_estim.py works fine in the
current setup: the resource "computation.pyx" utilises prange, and the
class EstimProcedure is invoked by setting "paralell=False", which means
that "mle_estimation.pyx" does not call
multiprocessing.Pool. Notice that the python notebook
"explore_saved_test_estim.ipynb" shows good results for the estimation
conducted on 2020-04-03 at 10:34. This is the first building block of
the eventual presentation of the estimation procedures.