model.initialise_from_partial (_calibration)
Previously, in "model.pyx" the function "initialise_from_partial" assumed
that the user was working with calibration on empirica data. This was
not convenient when testing test_estim_multijob.py. Hence, a distinction
was made between "model.initialise_from_partial_calibration" and
"model.initialise_from_partial". The former inherits the previous version
of the function, and it assumes the calibration on empirical data.
Instead, the latter (model.initialise_from_partial) does not assume that
and performs less task but its usage is more generic.
 On branch master
 Your branch is up-to-date with 'origin/master'.

 Changes to be committed:
	modified:   calibration/main.py
	modified:   calibration/main_EliteBook.py
	modified:   sdhawkes_powerlaw/modelling/model.pyx
	modified:   tests/test_estim_multijob.py