New way to merge after partial
Previously, the function "initialise_from_partial" in the class
"SDHawkes" called the functions "set_hawkes_parameters". This caused the
module to overwrite possible pre-existing values of the parameters. Now,
this function call has been removed.
Moreover, the estimation of transition probabilities and of dirichlet
parameters have been moved to the class
"mle_estiomation.EstimProcedure". This is true also for
"initialise_from_partial_calibration".
 Changes to be committed:
	modified:   sdhawkes_powerlaw/modelling/model.pyx
	modified:   sdhawkes_powerlaw/resources/mle_estimation.pyx