New multijob routine for test
Previously, the multijob routine for tests was conceived to be used on
the imperial college hpc cluster only. This meant in particular that it
could not be tested locally on my laptop. Exploiting the job control
flow on linux, a similar routine can be emulated locally. This commit
updates the script used to implement such routine. Note also the
modification in tests/test_estim_multijob.py regarding the boolean
variable "switch" in the function "estimate".
 On branch master
 Your branch is up-to-date with 'origin/master'.

 Changes to be committed:
	modified:   sdhawkes_powerlaw/modelling/model.pyx
	renamed:    tests/launch_multijob.sh -> tests/launch_hpc-cluster.sh
	new file:   tests/launch_mj.sh
	modified:   tests/test_estim_multijob.py