New way to call MinimisationProcedure
Because of the routine in batches, the minimisation procedure involves
producing the numpy arrays 'labelled_times' and 'count' for every batch;
so they cannot be computed beforehand and passed to the procedure as
done when no batches are used. This commit stores the changes made to
the resource "mle_Estimation.pyx" and to "model.pyx" in order to modify
the way in which they call and initialise the class MinimisationProcedure
of the resource "minimisation_algo.pyx".
 On branch minim_batch
 Changes to be committed:
	modified:   sdhawkes_powerlaw/modelling/model.pyx
	modified:   sdhawkes_powerlaw/resources/minimisation_algo.pyx
	modified:   sdhawkes_powerlaw/resources/mle_estimation.pyx
	modified:   tests/test_estim_multijob.py