max_base_rate in minimisation_algo.pyx
It has been noticed that the minimisation algorithm used in the gradient
descent tends to overestimate the base rates of the proint process. The
estimated values of the base rates were found to be even 10 times
greater than the true values. To correct this, an upperbound for these
values is forced into the minimisation procedure. The upperbound is
determined by a crude estimate on the average intensity observed in the
sample.

 On branch master
 Your branch is up-to-date with 'origin/master'.

 Changes to be committed:
	modified:   sdhawkes_powerlaw/resources/computation.pyx
	modified:   sdhawkes_powerlaw/resources/minimisation_algo.pyx