BOOKMARK: perf improvements in computation.pyx
This commit represents an important improvement in the performance of
some of the functions in the resource "computation.pyx". Some for-loops
have been nested together to improve the evaluation time of the
loglikelihood of the point process. Moreover, the resource
"computation.pyx" now offers the alternative whther to use or not the
function "prange" in the module cython.parallel. The gain in performance
are recorded in the python notebook
tests/performance_measurements/performance_notebook.ipynb. Notice
however, that "prange" is incompatible with multiprocessing.Pool and
user will have to decide whether to use one or the other (or neither).
This done via the boolean variables "use_prange" and "parallel", the
former controlling whether to use "prange", the latter controlling
whether to use "multiprocessing.pool".

 On branch master
 Your branch is up-to-date with 'origin/master'.

 Changes to be committed:
	modified:   sdhawkes_powerlaw/modelling/model.pyx
	modified:   sdhawkes_powerlaw/resources/computation.pyx
	modified:   sdhawkes_powerlaw/resources/minimisation_algo.pyx
	modified:   sdhawkes_powerlaw/resources/mle_estimation.pyx
	deleted:    sdhawkes_powerlaw/resources/nonparam_estimation.cpython-37m-x86_64-linux-gnu.so
	modified:   sdhawkes_powerlaw/resources/nonparam_estimation.pyx
	deleted:    sdhawkes_powerlaw/resources/simulation.cpython-37m-x86_64-linux-gnu.so
	modified:   sdhawkes_powerlaw/resources/simulation.pyx
	modified:   setup.py
	new file:   tests/performance_measurements/measure_exectime.cpython-37m-x86_64-linux-gnu.so
	new file:   tests/performance_measurements/measure_exectime.pyx
	new file:   tests/performance_measurements/performance_notebook.ipynb
	new file:   tests/performance_measurements/test_perf.py
	modified:   tests/test_estim_multijob.py