Calibration INTC_2019-01-04_36000-39000
Notice the usage of the PBS command
 PBS -l select=1:cpus=24:mem=32gb:mpiprocs=1:ompthreads=24
to specify the number of threads to use when "prange" is called.
Also, notice that the option "-march=native" must be removed from the
file setup.py, because otherwise the c files are compied using an
optimiser that can result in the error "illegal instruction".

 Changes to be committed:
   (use "git reset HEAD <file>..." to unstage)

	modified:   calibration/launch.sh
	modified:   calibration/main.py
	modified:   calibration/wrapper_submission_mle.sh
	modified:   sdhawkes_powerlaw/resources/mle_estimation.pyx
	modified:   setup.py